+++ contracts/token/BasicToken.sol
@@ -10,14 +10,14 @@ import '../SafeMath.sol';
  * @dev Basic version of StandardToken, with no allowances. 
  */
 contract BasicToken is ERC20Basic {
-  using SafeMath for uint;
+  using SafeMath for uint256;
 
-  mapping(address => uint) balances;
+  mapping(address => uint256) balances;
 
   /**
    * @dev Fix for the ERC20 short address attack.
    */
-  modifier onlyPayloadSize(uint size) {
+  modifier onlyPayloadSize(uint256 size) {
      if(msg.data.length < size + 4) {
        throw;
      }
@@ -29,7 +29,7 @@ contract BasicToken is ERC20Basic {
   * @param _to The address to transfer to.
   * @param _value The amount to be transferred.
   */
-  function transfer(address _to, uint _value) onlyPayloadSize(2 * 32) {
+  function transfer(address _to, uint256 _value) onlyPayloadSize(2 * 32) {
     balances[msg.sender] = balances[msg.sender].sub(_value);
     balances[_to] = balances[_to].add(_value);
     Transfer(msg.sender, _to, _value);
@@ -38,9 +38,9 @@ contract BasicToken is ERC20Basic {
   /**
   * @dev Gets the balance of the specified address.
   * @param _owner The address to query the the balance of. 
-  * @return An uint representing the amount owned by the passed address.
+  * @return An uint256 representing the amount owned by the passed address.
   */
-  function balanceOf(address _owner) constant returns (uint balance) {
+  function balanceOf(address _owner) constant returns (uint256 balance) {
     return balances[_owner];
   }
